# solution/backend/pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "PROD Backend 2025 Advertising Platform API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/ads": {
            "get": {
                "description": "Возвращает информацию о рекламе, соответствующей таргетингу клиента.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Реклама"
                ],
                "summary": "Получить информацию о рекламе для клиента",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Идентификатор клиента",
                        "name": "client_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о кампании",
                        "schema": {
                            "$ref": "#/definitions/model.AdsResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный идентификатор клиента",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Клиент не найден или не найдена подходящая кампания",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ads/{campaignID}/click": {
            "post": {
                "description": "Регистрирует клик клиента по рекламе, проверяя наличие предыдущих действий (просмотр и клик) и сроки действия кампании",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Реклама"
                ],
                "summary": "Зарегистрировать клик по рекламе",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Идентификатор кампании",
                        "name": "campaignID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Идентификатор клиента",
                        "name": "ClientID",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AdsRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Клик успешно зарегистрирован"
                    },
                    "400": {
                        "description": "Неверный формат идентификаторов клиента или кампании, или клиент уже кликнул",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Клиент или кампания не найдены",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "410": {
                        "description": "Кампания не началась или завершена",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/advertisers/bulk": {
            "post": {
                "description": "Создает или обновляет информацию о рекламодателе с проверкой уникальности идентификатора",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Рекламодатели"
                ],
                "summary": "Создать или обновить рекламодателя",
                "parameters": [
                    {
                        "description": "Список рекламодателей для создания или обновления",
                        "name": "advertisers",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Advertiser"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Информация о рекламодателях, которые были успешно созданы или обновлены",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Advertiser"
                            }
                        }
                    },
                    "400": {
                        "description": "Невалидные данные или повторяющиеся идентификаторы рекламодателей",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/advertisers/{advertiserID}": {
            "get": {
                "description": "Возвращает информацию о рекламодателе по указанному идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Рекламодатели"
                ],
                "summary": "Получить информацию о рекламодателе по ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Идентификатор рекламодателя",
                        "name": "advertiserID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о рекламодателе",
                        "schema": {
                            "$ref": "#/definitions/model.Advertiser"
                        }
                    },
                    "400": {
                        "description": "Неверный формат идентификатора рекламодателя",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Рекламодатель не найден",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.AdsRequest": {
            "type": "object",
            "required": [
                "client_id"
            ],
            "properties": {
                "client_id": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        },
        "model.AdsResponse": {
            "type": "object",
            "required": [
                "ad_id",
                "ad_text",
                "ad_title",
                "advertiser_id"
            ],
            "properties": {
                "ad_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "ad_text": {
                    "type": "string"
                },
                "ad_title": {
                    "type": "string"
                },
                "advertiser_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "image_url": {
                    "type": "string",
                    "format": "url",
                    "example": "https://domain.com/image.jpg"
                }
            }
        },
        "model.Advertiser": {
            "type": "object",
            "required": [
                "advertiser_id",
                "name"
            ],
            "properties": {
                "advertiser_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.ErrorResponse": {
            "type": "object",
            "required": [
                "message",
                "status"
            ],
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}